Make more tests pass ...

+++ gpaw/wavefunctions/base.py
@@ -409,7 +409,7 @@ class WaveFunctions(EmptyWaveFunctions):
 
         if rank == 0:
             # allocate full wave function and receive
-            psit_G = self.empty(dtype=self.dtype, global_array=True,
+            psit_G = self.empty(global_array=True,
                                 realspace=realspace)
             world_rank = (kpt_rank * self.gd.comm.size *
                           self.band_comm.size +
